refactor: VO적용
