+++ src/main/java/com/ssg/martgowmsfullstack/service/RentServiceImpl.java
@@ -1,6 +1,7 @@
 package com.ssg.martgowmsfullstack.service;
 
 
+import com.ssg.martgowmsfullstack.domain.RentHistoryVO;
 import com.ssg.martgowmsfullstack.dto.CostInfoDTO;
 import com.ssg.martgowmsfullstack.dto.RentHistoryDTO;
 import com.ssg.martgowmsfullstack.dto.SectorDTO;
@@ -8,17 +9,14 @@ import com.ssg.martgowmsfullstack.dto.WarehouseDTO;
 import com.ssg.martgowmsfullstack.mapper.RentMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.ui.Model;
-
-import java.sql.Date;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+import java.util.stream.Collectors;
+
+
 
 @Service
 @Log4j2
@@ -27,16 +25,19 @@ import java.util.Map;
 public class RentServiceImpl implements RentService {
 
     @Autowired
-    RentMapper rentMapper;
+    private final RentMapper rentMapper;
 
+    @Autowired
+    private final ModelMapper modelMapper;
 
     public List<WarehouseDTO> getAllWarehouses() {
         List<WarehouseDTO> warehouses = rentMapper.getAllWarehouses();
-        for (WarehouseDTO warehouse : warehouses) {
+        warehouses.forEach(warehouse -> {
             warehouse.setStatus(getSectorStatus(warehouse.getWarehouseId()));
-        }
+        });
         return warehouses;
     }
+
     public String getSectorStatus(int warehouseId) {
         List<String> sectorStatuses = rentMapper.getSectorStatus(warehouseId);
         boolean anyAvailable = sectorStatuses.stream()
@@ -45,34 +46,47 @@ public class RentServiceImpl implements RentService {
     }
 
     public List<SectorDTO> getAllSector(int warehouseId) {
-        return rentMapper.getAllSectors(warehouseId);
+        List<SectorDTO> vo = rentMapper.getAllSectors(warehouseId);
+        return vo.stream()
+                .map(i -> modelMapper.map(i,SectorDTO.class))
+                .collect(Collectors.toList());
     }
 
     public List<CostInfoDTO> getAllCostInfo(int wareHouseId, String sectorId) {
-        return rentMapper.getCostInfo(wareHouseId,sectorId);
+        List<CostInfoDTO> vo = rentMapper.getCostInfo(wareHouseId, sectorId);
+        return vo.stream()
+                .map(i -> modelMapper.map(i, CostInfoDTO.class))
+                .collect(Collectors.toList());
     }
 
 
     public void saveRentHistory(RentHistoryDTO rentHistoryDTO) {
-        rentMapper.saveDb(rentHistoryDTO);
+        RentHistoryVO rentHistoryVO = modelMapper.map(rentHistoryDTO, RentHistoryVO.class);
+        rentMapper.saveDb(rentHistoryVO);
     }
 
     public List<RentHistoryDTO> holdRentList(String adminId) {
-       return rentMapper.getHoldRentHistory(adminId);
+        List<RentHistoryDTO> vo = rentMapper.getHoldRentHistory(adminId);
+        return vo.stream()
+                .map(i -> modelMapper.map(i, RentHistoryDTO.class))
+                .collect(Collectors.toList());
     }
 
     public List<RentHistoryDTO> inProgressRentList(String adminId) {
-        return rentMapper.getInProgressRentHistory(adminId);
+        List<RentHistoryDTO> vo = rentMapper.getInProgressRentHistory(adminId);
+        return vo.stream()
+                .map(i -> modelMapper.map(i, RentHistoryDTO.class))
+                .collect(Collectors.toList());
     }
 
-   public void approveRentHistory(int rentNum, String adminId){
-        rentMapper.updateAdminId(rentNum,adminId);
+    public void approveRentHistory(int rentNum, String adminId) {
+        rentMapper.updateAdminId(rentNum, adminId);
         rentMapper.updateUserAdminId();
 
-   }
+    }
 
-    public void confirmRentHistory(int rentNum, String adminId){
-        rentMapper.completedRentStatus(rentNum,adminId);
+    public void confirmRentHistory(int rentNum, String adminId) {
+        rentMapper.completedRentStatus(rentNum, adminId);
     }
 
 }
