+++ src/main/webapp/WEB-INF/views/pages-rent-last.jsp
@@ -61,7 +61,9 @@
             <div class="navbar-collapse collapse">
                 <ul class="navbar-nav navbar-align ms-auto">
                     <li class="nav-item">
+                        <a class="nav-link" href="${pageContext.request.contextPath}/user/mypage">
                         <span class="nav-link"><i class="fas fa-user-circle"></i> <%= user.getUsername() %>님</span>
+                        </a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" href="${pageContext.request.contextPath}/logout">
