+++ src/main/webapp/WEB-INF/views/pages-rent-warehouse.jsp
@@ -60,7 +60,9 @@
             <div class="navbar-collapse collapse">
                 <ul class="navbar-nav navbar-align ms-auto">
                     <li class="nav-item">
-                        <span class="nav-link"><i class="fas fa-user-circle"></i> <%= user.getUsername() %>님</span>
+                        <a class="nav-link" href="${pageContext.request.contextPath}/user/mypage">
+                            <i class="fas fa-user-circle"></i> <%= user.getUsername() %>님
+                        </a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" href="${pageContext.request.contextPath}/logout">
